Context: impl From<ContextConfig> for Context
